frontend(components): moved nav and models into a folder
