+++ frontend/src/components/models/DirectoryInfoModal.tsx
