+++ frontend/src/components/navbar/NavCard.tsx
