+++ frontend/src/components/navbar/VerticalStepper.module.css
