+++ frontend/src/components/navbar/VerticalStepper.tsx
